New Data

+++ queuing_server/src/core/constants.rs
@@ -1,3 +1,3 @@
 pub const MAX_CUSTOMERS_IN_QUEUE: u8 = 5;
-pub const MAX_NUMBER_OF_SERVERS: u8 = 5;
+pub const MAX_NUMBER_OF_SERVERS: usize = 5;
 pub const OPEN: bool = true;
